Refactored out iterator tests into own file. Added example
